Make XMLHttpRequest override-friendly to enable use with Angular 2
The new implemetnation appears to behave like a regular prototype-based
constructor, making it possible for Zone.js to patch it. Also see
angular/zone.js#335.

The previous implementation could not be overridden by Angular's Zone.js
because it broke the prototype chain in different ways, depending on
enablement of the CORS proxy.